+++ apps/nextjs-app/src/_backend/api/rest/post-repository.ssr.ts
@@ -1,7 +1,7 @@
 import { HttpInternalServerError, HttpNotFound } from '@belgattitude/http-exception';
+import type { UnPromisify } from '@teable-group/core';
+import { assertNonEmptyString } from '@teable-group/core';
 import type { PrismaClientDbMain } from '@teable-group/db-main-prisma';
-import type { UnPromisify } from '@teable-group/ts-utils';
-import { assertNonEmptyString } from '@teable-group/ts-utils';
 
 export type IGetPosts = UnPromisify<ReturnType<typeof PostRepositorySsr['prototype']['getPosts']>>;