+++ apps/nextjs-app/src/_backend/config/prisma/db-main-prisma.config.ts
@@ -1,5 +1,5 @@
+import { assertNonEmptyString } from '@teable-group/core';
 import { PrismaManager, PrismaClientDbMain } from '@teable-group/db-main-prisma';
-import { assertNonEmptyString } from '@teable-group/ts-utils';
 
 const isDev = process.env?.NODE_ENV === 'development';