+++ apps/nextjs-app/src/_backend/features/poem/SearchPoems/SearchPoems.types.ts
@@ -1,4 +1,4 @@
-import type { UnPromisify } from '@teable-group/ts-utils';
+import type { UnPromisify } from '@teable-group/core';
 import type { SearchPoemsQuery } from './SearchPoemsQuery';
 
 export interface ISearchPoemsParams {