+++ apps/nextjs-app/src/pages/api/rest/post/[id].ts
@@ -1,7 +1,7 @@
 import { HttpBadRequest, HttpMethodNotAllowed } from '@belgattitude/http-exception';
+import { assertSafeInteger, stringToSafeInteger } from '@teable-group/core';
 import { JsonApiResponseFactory } from '@teable-group/sdk/api/json-api';
 import { JsonApiErrorFactory } from '@teable-group/sdk/api/json-api/json-api-error.factory';
-import { assertSafeInteger, stringToSafeInteger } from '@teable-group/ts-utils';
 import type { NextApiRequest, NextApiResponse } from 'next';
 import { PostRepositorySsr } from '@/_backend/api/rest/post-repository.ssr';
 import { prismaClient } from '@/_backend/config/container.config';