+++ packages/core/src/convert/string-convert.ts
@@ -8,15 +8,9 @@ export function stringToSafeInteger(value: string | unknown): number | null {
 }
 
 export function stringToFloat(value: string | unknown): number | null {
-  if (
-    !isParsableNumeric(
-      typeof value === 'number' ? value.toString(10) : value ?? ''
-    )
-  ) {
+  if (!isParsableNumeric(typeof value === 'number' ? value.toString(10) : value ?? '')) {
     return null;
   }
-  const v = Number.parseFloat(
-    typeof value === 'string' ? value : (value as number).toString(10)
-  );
+  const v = Number.parseFloat(typeof value === 'string' ? value : (value as number).toString(10));
   return Number.isNaN(v) ? null : v;
 }