+++ packages/sdk/src/api/json-api/json-api-response.factory.ts
@@ -1,4 +1,4 @@
-import { isPlainObject } from '@teable-group/ts-utils';
+import { isPlainObject } from '@teable-group/core';
 import type {
   IJsonApiError,
   IJsonApiErrorResponse,