+++ packages/sdk/src/api/json-api/json-api.typeguard.ts
@@ -1,13 +1,11 @@
-import { isPlainObject } from '@teable-group/ts-utils';
+import { isPlainObject } from '@teable-group/core';
 import type {
   IJsonApiErrorResponse,
   IJsonApiResponse,
   IJsonApiSuccessResponse,
 } from './json-api-response.types';
 
-export const isJsonApiResponse = <T = unknown>(
-  val: unknown
-): val is IJsonApiResponse<T> => {
+export const isJsonApiResponse = <T = unknown>(val: unknown): val is IJsonApiResponse<T> => {
   return isPlainObject(val) && typeof val?.success === 'boolean';
 };
 
@@ -17,13 +15,8 @@ export const isJsonApiSuccessResponse = <T = unknown>(
   return isJsonApiResponse<T>(val) && val.success && 'data' in val;
 };
 
-export const isJsonApiErrorResponse = (
-  val: unknown
-): val is IJsonApiErrorResponse => {
+export const isJsonApiErrorResponse = (val: unknown): val is IJsonApiErrorResponse => {
   return (
-    isJsonApiResponse<unknown>(val) &&
-    !val.success &&
-    'errors' in val &&
-    Array.isArray(val.errors)
+    isJsonApiResponse<unknown>(val) && !val.success && 'errors' in val && Array.isArray(val.errors)
   );
 };