+++ packages/sdk/src/context/table.ts
@@ -1,5 +1,5 @@
 import React from 'react';
-import type { Table } from '@teable-group/ts-utils';
+import type { Table } from '@teable-group/core';
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
 export const TableContext = React.createContext<Table | null>(null);