+++ packages/core/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @teable-group/ts-utils
+# @teable-group/core
 
 ## 1.1.1
 
@@ -13,6 +13,6 @@
 
 ### Minor Changes
 
-- [#2713](https://github.com/teable-group/teable/pull/2713) [`e279d984`](https://github.com/teable-group/teable/commit/e279d984d01937264b077a111863b38d15fcb1d0) Thanks [@belgattitude](https://github.com/belgattitude)! - Move typescript utils to @teable-group/ts-utils
+- [#2713](https://github.com/teable-group/teable/pull/2713) [`e279d984`](https://github.com/teable-group/teable/commit/e279d984d01937264b077a111863b38d15fcb1d0) Thanks [@belgattitude](https://github.com/belgattitude)! - Move typescript utils to @teable-group/core
 
-- [#2713](https://github.com/teable-group/teable/pull/2713) [`e279d984`](https://github.com/teable-group/teable/commit/e279d984d01937264b077a111863b38d15fcb1d0) Thanks [@belgattitude](https://github.com/belgattitude)! - Example ts-utils packages as esm with jest and vitest
+- [#2713](https://github.com/teable-group/teable/pull/2713) [`e279d984`](https://github.com/teable-group/teable/commit/e279d984d01937264b077a111863b38d15fcb1d0) Thanks [@belgattitude](https://github.com/belgattitude)! - Example core packages as esm with jest and vitest